+++ src/main/java/subway/application/service/line/LineCommandService.java
@@ -11,7 +11,7 @@ import subway.application.port.out.line.LoadLinePort;
 import subway.application.port.out.line.PersistLinePort;
 import subway.common.exception.NoSuchLineException;
 import subway.common.exception.SubwayIllegalArgumentException;
-import subway.domain.LineInfo;
+import subway.domain.line.LineInfo;
 
 @Service
 @Transactional
