+++ src/main/java/subway/common/mapper/SectionMapper.java
@@ -1,7 +1,7 @@
 package subway.common.mapper;
 
 import subway.adapter.out.persistence.entity.SectionEntity;
-import subway.domain.Section;
+import subway.domain.section.Section;
 
 public class SectionMapper {
