+++ src/main/java/subway/domain/route/Route.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 import subway.common.exception.SubwayIllegalArgumentException;
-import subway.domain.Line;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.station.Station;
 
 public class Route {
