+++ src/main/java/subway/domain/section/Sections.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.section;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import subway.common.exception.SubwayIllegalArgumentException;
+import subway.domain.station.Station;
 
 public class Sections {
