+++ src/test/java/subway/application/service/line/LineQueryServiceTest.java
@@ -16,8 +16,8 @@ import org.junit.jupiter.api.Test;
 import subway.application.port.in.line.dto.response.LineQueryResponse;
 import subway.application.port.out.line.LoadLinePort;
 import subway.common.exception.NoSuchLineException;
-import subway.domain.Line;
-import subway.domain.Section;
+import subway.domain.line.Line;
+import subway.domain.section.Section;
 import subway.fixture.SectionFixture.이호선_삼성_잠실_2;
 import subway.fixture.SectionFixture.이호선_역삼_삼성_3;
 import subway.fixture.StationFixture.건대역;
