+++ src/test/java/subway/application/service/route/RouteQueryServiceTest.java
@@ -20,9 +20,9 @@ import subway.application.port.out.line.LoadLinePort;
 import subway.application.port.out.route.RouteFinderPort;
 import subway.application.port.out.station.LoadStationPort;
 import subway.common.exception.NoSuchStationException;
-import subway.domain.Line;
-import subway.domain.Section;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.section.Section;
+import subway.domain.station.Station;
 import subway.domain.fare.Fare;
 import subway.domain.fare.FarePolicy;
 import subway.domain.route.Route;
