+++ src/test/java/subway/domain/fare/LineSurchargePolicyTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import subway.domain.Line;
-import subway.domain.Section;
+import subway.domain.line.Line;
+import subway.domain.section.Section;
 import subway.domain.route.Route;
 import subway.domain.route.RouteSection;
 import subway.fixture.StationFixture.삼성역;
