+++ src/test/java/subway/fixture/SectionFixture.java
@@ -4,8 +4,8 @@ import subway.adapter.in.web.section.dto.AddStationToLineRequest;
 import subway.adapter.out.persistence.entity.LineEntity;
 import subway.adapter.out.persistence.entity.SectionEntity;
 import subway.adapter.out.persistence.entity.StationEntity;
-import subway.domain.Section;
-import subway.domain.Station;
+import subway.domain.section.Section;
+import subway.domain.station.Station;
 import subway.fixture.LineFixture.삼호선;
 import subway.fixture.LineFixture.이호선;
 import subway.fixture.StationFixture.건대역;
