+++ src/test/java/subway/integration/RouteIntegrationTest.java
@@ -11,10 +11,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import subway.adapter.out.persistence.LineJdbcRepository;
 import subway.adapter.out.persistence.StationJdbcRepository;
 import subway.application.port.in.route.dto.response.RouteQueryResponse;
-import subway.domain.Line;
-import subway.domain.LineInfo;
-import subway.domain.Section;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.line.LineInfo;
+import subway.domain.section.Section;
+import subway.domain.station.Station;
 import subway.fixture.StationFixture.강남역;
 import subway.fixture.StationFixture.고터역;
 import subway.fixture.StationFixture.교대역;
